Improve `useSubscribeToContractEvent` hook
We can't add `indexedFilterParams` array as dependency in `useEffect`
hook because this variable will always point to a new array on every
render. Because we know that an event in solidity can have up to 3
indexed parameters, we can extract the first three parameters from array
and use them in `useEffect` dependency array. Thanks to this the
`useSubscribeToContractEvent` hook will be triggered when one of these
params changes instead of on every render.