Deploy a Multicall contract to local network
This commit adds hardhat project in order to deploy the `Multicall`[1]
contract to local network and test the integration with this contract in
T dapp. We copied the `Multicall` contract[2] because we want to use
hardhat to simplify the deployment process. This hardhat project is for
development purposes only- there is no need to deploy this contract to
`Mainnet` or `Ropsten` network because we can use already deployed
contracts[3].

Ref:
[1]: https://github.com/makerdao/multicall
[2]: https://github.com/makerdao/multicall/blob/master/src/Multicall.sol
[3]: https://github.com/makerdao/multicall#multicall-contract-addresses