Refactor InfoBox component
Build this component in chakra way so we can get rid of the `useModal`
hook in this component. Thanks to this the `InfoBox` component will not
change style if the modal is open and is not used inside `Modal`
component. The previous implementation changes the `InfoBox` color if
the modal is open even the component was outside the `Modal` component.
This commit will fix it.