Fix fetching token balances
This is a workaround because if we provide wrong param to the contrct's
function eg. `undefined` the
`_.contract?.interface.encodeFunctionData(_.method, _.args),` will fail.
We want to fetch token balances only if a user is connected to a wallet-
to avoid passing the `undefined` if a user is not connected yet we pass
the `AddressZero` to the `useTokensBalanceCall` hook which uses the
`useMulticall` hook.