Deploy to GCP bucket after manual dispatch of workflow
So far the step publishing the T dashboard to
`dashboard.test.threshold.network` GCP bucket was executed only when new
changes were being introduced to `main` branch.
There may be cases when we want to re-deploy T dashboard on Testnet
without pushing anything to `main`. This may happen when there's a
change in the dependecncies of the project that we want to see reflected
in the dashboard. To make such deploy possible, we're allowing for
execution of `Deploy to GCP bucket` step for 'workflow_dispatch' event.