Update T dashboard code to use Goerli testnet
As part of the task of switching T dApp from Ropsten to Goerli testnet
(caused by announced soon deprecation of Ropsten) we need to update a
bunch of maps listing supported networks and need to point to some
contracts deployed on Goerli network.

Note: As NuCypher hasn't deployed the `SimplePreApplication` contract on
Goerli testnet, we're using stub contract.