Update Application interface
Calculate the deauthorization request date if possible and check if the
deauthorization request is in an active stake. If a
`remainingAuthorizationDecreaseDelay` is equal `MAX_UINT64` the
deauthorization reqest is pending and an operator have to call
`joinSortitionPool` or `updateOperatorStatus` to activate the request.
In that case we can't estimate when the deauthorization request started.
Also we can't estimate this date when the
`_remainingAuthorizationDecreaseDelay` is equal `0`.