Display deauth completed modal correctly
We should display this modal only if the `AuthorizationDecreaseApproved`
is related to the currently logged account. We can assume if stake for a
given staking provider(from event) exists, this event is related to the
currently logged account because we fetch the app data based on the
already fetched stakes.