Update `StakeApplications` component
The alert saying the tbtc and random beacon earn bundled rewards display
at the beginning for 1 seconds after connecting to a wallet. It happends
before the data is fetching and the `isAuthorized` property is set to
`true` as default. Adding additional condition that fetching status
should be `falsy` for both apps did the trick.