Update the `StakeCard` component
Disable `Top-up/Unstake` button when the currently connected user is not
an owner of this stake. When an authorizer goes to the
`staking/<stakingProvider>/authorize` page we store the stake data and
staking app data in the same place in redux store as for an owner so
when an authrozier then goes to the `staking` page will see stake card
with a stake but that user can't do any action because is not an owner
of the stake.