Fix display operator mapping effect
There was a bug where when you close the login modal too quick, before
`setStakes` action was dispatched, then the effect got blocked on `await
listenerApi.condition` and was waiting for another action to dispatch where
there were not any at this moment.

To fix that I've added a condition to check if login modal was closed, if not,
then we wait with the `condition()` function for it to be closed. If it was
closed then we don't need `condition()` since this effect was run on `setStakes`
action so we have all data fetched already.