Add posthog hooks
For the posthog to work properly, we need to manually capture the
`$pageview` event because posthog sends the `pageview` whenever it
loaded. For one-page app like pur T dapp, that means it'll only send a
pageview once, when the app loads. This commit adds support for
capturing the `$pageview` whenever the `location`(url) changes.

Also we want to identify the user by the eth address so we need to reset
the posthog once the account changes and identify user by the new
account. The reset is also required if the user logged out.