Improvde getStakingProviderOperatorInfo effect
Before calling `rolesOf` method from the contract we want to check if there are
stakes in `staking` slice that have the owner === staking provider (since we
keep them there form the owner perspective, so owner === currently connected
account). If yes then we get the data from there and we don't have to do a call
to contract.

Note: Even though the `getStakingProviderOperatorInfo` effect is called on
`setStakes` action, the store does not containe the stakes yet. Forutnately we
can get them from the `action` argument.