Improve Bridge page layout
We changed the grid layout to a stack layout in a Bridge page which will
eliminate some of the UI bugs caused by the Grid:
- too long TBTC balance card on Step 2 (because of the grid two rows on the
Bridge page were equal and thus it prolonged the TBTC balance card)
- (additionally) improved the UI looks on the resolution lower than 1280 px
(xl). Right now we display the modal window that says that this page won't work
on mobile devices on a resolution with width less than 1280px. This change makes
the page responsive on all resolutions.

Also we added a truncate and a maxW property to a deposit address container.
This will keep the rows in `MakeDeposit` container same width in step 1 and step
2 (and all the steps, but step 2 was the problem here).