Create a wrapper for merkle drop contract
In the future we should create a contract wrapper for each contract the
threshold lib integrates to keep the same interface. In some cases
multiple services need the same contract data like deployment block or
address. So to avoid duplicating the creation of each contract in each
service can create one contract instance and pass to the service in
constructor.