Fix tbtc-v2 utils
To install the `tbtc-v2.ts` dependency we use the `development` tag so
during the CI build the latest version of this library will always be
pulled. In this [PR](https://github.com/keep-network/tbtc-v2/pull/419)
the `DepositScriptParameters` interface has been changed and this commit
will adjust a utils function in T-dapp to these changes.