Revert "Fix tbtc-v2 utils"
This reverts commit 48aabf51ac531d05b39ce04fa1974f37b2ad9f88.

This would require bumping the `tbtc-v2.ts` dependency to the latest
version with `development` tag but it does bump other dependencies but
we do not want to do that on the release branch. In that case we are
going to set the exact version of the `tbtc-v2.ts`(we can get the exact
version from the `yarn.lock` file) dependency so CI
builds will succeed.