Get rid of MockTBTC class
We decided to remove the MockTBTC class and pass the mocked client (or real
client based on the flag [to do]) inside the `ThresholdConfig`. This way we
won't need a whole new interface for the TBTC wrapper.

Besides that there was and error with the compatibility of a `Signer` type
needed to create a Bridge instance. I've fixed that by bumping up the `ethers`
version to `5.5.2`.