Update the `Minting Card`
Based on the Figma views the link to the `Bridge` contract address
should be on every step and at the bottom of the card. In previous impl,
each step comoponent had link to the contract but it's unnecessary. We
can render this link in `MintingFlowRouter` component. This commit
refactors this.