Fix failing CI jobs
As described in #340, here we use resolution
(https://classic.yarnpkg.com/en/docs/selective-version-resolutions/) for
`@tenderly/hardhat-tenderly` dependency.  Without this, the command

```
yarn upgrade \
 @threshold-network/solidity-contracts@dapp-development-goerli \
 @keep-network/keep-core@1.8.1-goerli.0 \
 @keep-network/keep-ecdsa@goerli \
 @keep-network/tbtc@goerli \
 @keep-network/coverage-pools@goerli \
 @keep-network/ecdsa@dapp-development-goerli \
 @keep-network/random-beacon@dapp-development-goerli \
 @keep-network/tbtc-v2@dapp-development-goerli
```
is failing with the following error:

```
error tslog@4.7.1: The engine "node" is incompatible with this module.
Expected version ">=16". Got "14.21.1"
```

We know that such error occurs when one of the dependencies uses
`@tenderly/hardhat-tenderly` dependency in versions `>=1.4.x`, as
`1.4.0` introduced a dependency to `tslog` module which is incompatible
with Node.js v14. We updated our `development` packages to not use
`@tenderly/hardhat-tenderly` in versions `>=1.4.x`, but we haven't done
that yet for `goerli`/`dapp-development-goerli` packages. This is why we
need to force usage of the lower versions using the resolutions
functionality.