Fix rejecting reveal tx bug
There were two bugs when rejecting a reveal transaction:
- the modal with the error was not displayed
- there was a warning in the console that said that a non-seriazible value was
found in the state

The first bug I've fixed with checking if the transaction hash is returned
before using `closeModal()` function.

The second bug I've fixed with changing the way how we pass the error to the
modal. We passed a whole error from catch block which had an unknown type. We
now only pass the message. In the next commits I'll refactor it to have error
description and title passed.