Use `@keep-network/coverage-pools@development` dependency
In a commit `11a4adf950b3e68b442c6ca3a64dfbfc93e186e8` we made a temporary change
that forced using of a fixed version of the `@keep-network/coverage-pools`
dependency (`@1.1.0-dev.7`). This was because using newer (`development`-tagged
versions was resulting in the `expected manifest` error during `yarn upgrade
@keep-network/coverage-pools`. We didn't want to block the dashboard deployment
by this issue, so we decided to use latest working version until we figure out
the root cause of the error.
Now we are bringing back the `@keep-network/coverage-pools@development`
dependency, as we fixed the problem occurring for some of the packages and we've
published a new `development`-tagged `coverage-pools` package. The problem that
we think was causing the `expected manifest` issue was not using the `--exact`
flag when resolving dependencies during building of the `coverage-pools` package.
Due to that the versions of the updated dependencies in that package were being
resolved by adding `^`, which can cause problems when `-dev`, `-goerli`,
`-mainnet`, etc. suffixes are also present in the NPM registry. See
https://github.com/keep-network/coverage-pools/pull/230 for more info.