Fix utxo fetch in `MakeDeposit` component
We want to look for the utxos periodically so we are creating an interval inside
a useEffect. The `findAllUnspentTransactionOutputs` will fire immidietely and
then fire periodically every 10 seconds.

We also remove some unencessery code like calculating the deposit address again
from the deposit data. We store the deposit address inside a redux store so we
don't need that.