Create mapping for bitcoin address
There is a difference in namings - in TBTC we use main for mainnet and
bitcoin-address-validation uses mainnet. That's why we create a mapping -
`tbtcLibNetworkMap` - that will handle changing the network name to the one
specified in tbtc-v2.ts lib.

We are also getting rid of `_getBitcoinNetworkForTBTCLib` and
`getBitcoinNetwork` and instead we will keep just `bitcoinNetwork` getter.