Create new modal
Add a modal that should be displayed when the user clicks `back` button
in Step 2 of the deposit form. This modal asks user if they want to
generate a new deposit address and informs them about the consequences.
If a user confirms generating a new deposit address will be redirected
to the Step 1 and deposit data will be removed from local storage. If a
user closes this modal it means they want to use already generated
deposit address.