Change the way we pass props to filerecovery modal
After we changed the Identifier type in tbtc-v2.ts (by adding additional
`equals` function) we encouter an error in the console saying:

`A non-serializable value was detected in the state, in the path:
modal.props.depositScriptParameters.depositor`.

It happens because when we open the modal we pass the parameters through the
state and in this case we are opening `TbtcRecoveryFileModalModal`. A
non-serializable value in this case is `equals` function that we passed with
`depositScriptParameters` property.

To fix that I've changed the way how we apss props there - we pass each
parameter on it's own instead of the whole `DepositScriptParameter` object as
one property. So now instead of passing `Identifier` type we pass just an eth
address and we change it to `Identifier` inside the modal using
`EthereumAddress.from(ethAddress)`.