Clear the interval when the UTXO is found
We don't want to encourage users to send multiple transactions to the deposit
address. That's why after we find any UTXO in the deposit address we are
stopping the 10-second interval that search for the UTXOs. We don't have to
search for them anymore even if the UTXO is revealed (TODO: in that case [if
it's revealed] we will navigate him back to step 1 and remove the data from
local storage).