Set correct amount to mint
In previous implementation the `tBTCMintAmount` was equal to BTC deposit
amount on Bitcoin chain and in Satoshi unit. Actually, the amount of
tBTC that will be minted is slighlty lower due to fees. Also the field
name indicates that this amount should be in the ERC20 standart- 18
decimals instead of Satoshi. This commit fixes that and takes into
account fees and sets the amount to mint in ERC20 standart.