Get the proper utxo
When there are multipe UTXOs sent to the deposit address the
`findAllUnspentTransactionOutputs` method (from `tbtc-v2.ts` lib) reverse the
order of the UTXOs, which means they are returned from the newest to oldest one.

To save the proper UTXO we use `pop()` method on the `utxos` array to get the
last element of it.