Integrate second track deposit verification with Sentry telemetry
Here we integrate the second track deposit address verification with deposit
telemetry. The verification calls a Cloud Function deployed
on `keep-test` and is done right before sending deposit data to Sentry.
The verification outcome is attached to the telemetry data as two additional
fields:
- `verificationStatus` which can have three different values
  - `valid`: verification service responded with the same address as the
  generated one
  - `invalid`: verification service responded with a different address
  - `error`: request or verification service error
- `verificationResponse` which holds the complete request outcome

The verification service availability does not impact the dApp functionality.
The deposit flow and telemetry works even if something is wrong on service side.