Make ETH address field readonly
The dApp allows user to create a deposit address using address other than his
own. We shouldn't allow this because after user sends his funds to the deposit
address he won't be able to reveal the deposit. It will throw an error `Wrong
32-byte script hash`.

We are adding two things:
- making eth address readonly,
- add a `throw error` in onSubmit function just in case someone manage to submit
the form with a different address.