Improve mocked bitcoin client
Changes:
1. Adds new methods - `_mockConfirmationsForTransaction`

The function mocks the confirmations for the given transaction based on it's
hash. Adds a new confirmation every 8 seconds. Additionally it can take a second
parameter which will define how many confirmations should be added to the tx.

2. Reverse the order of the UTXOs returned in `findAllUnspentTransactionOutputs`
function

The real electrum client in `tbtc-v2.ts` lib reverses the order of the UTXOs
returned from the `findAllUnspentTransactionOutputs` function. The mocked
bitcoin client should do that to.

3. Change `mockDepositTransaction` function name to `_mockDepositTransaction`
and make it private.

The function is used only inside of the class so we are renaming it to
`_mockDepositTransaction` to indicate that it's private.