Add effect for checking tx confirmations
Adds an effect that checks the tx confirmation for a given transaction
(based on a given th hash). The effect is fired after the utxo is found
and saved in the store. We check the number of confirmations every 10
seconds and stop the effect when the number of confirmations is 6 or
more, because this is the most that we would ever need for the UI.