Don't fire the effect if we have all confirmations
We don't need to fire the whole effect and call `getTransactionConfirmations`
method if we already have all neede number of confirmations. Because of that
I've introduced new method in TBTC class - `minimumNumberOfConfirmationsNeeded`.
Thanks to that we can get the minimum number of confirmations needed based on
the value of the transaction and we can use check in effect if we already have
minimum transactions required.