Improve the message for ui.click and clean code
UI click can happen on any html element, not only button so we are changin the
variable name form button to element (and all it's occurences in the code).
Besides that we are adding element type and id to the breadcrumb message so that
it will be even easier to debug the Sentry error.