Move window.ethereum type declaration
To the `reac-app-env.d.ts` file. The Coinbase Wallet and MetaMask inject
the provider to the same field `window.ethereum` but they have different
types so we can't have 2 global declaration for the same field. So here
we move the `window.etherreum` declaration to the `reac-app-env.d.ts`
file and set the type to `unknown` to force developers to use the exact
type of the provider when using `window.ethereum`.