Extract a common component
Extract a common list item component with outline variant as we use them
in several places in dapp. We can't add a new variant using chakra
theming because the `ListItem` is a part of the `List` component and
does not have separate theme. In that case we should pass the eg
`variant="otuline"` to the `List` component but in my opinion does not
fit here because list can be order or unordered.