Update the `findRefreshedKeepStakes` method
We should find all `OwnerRefreshed` events where the `owner` was set as
a new owner or old owner of the stake instead of only taking into
account the `newOwner` param from `OwnerRefreshed` event. For example,
the `Staked` event was emitted with the owner address that we are
looking for but then the owner was changed(`OwnerRefreshed` event
emitted with `oldOwner = owner`) meaning the owner is not the owner of
this staking- we should skip that address.