Add a new env variable
Add the new `GET_EVENTS_BLOCK_INTERVAL` env variable- defines a range of
blocks using in the partial events pulls mechanism. The main idea behind
passing this value as an env variable is that we can easily decrease the
range and redeploy the T dapp to prod w/o changing anything in code. The
default value is `10_000`- if this variable is omitted.