Create a Step component based on the TDS
Create a new `Step` component and refactor the `TimelineItem` component
in `Bridge` domain- build this component using the newly created `Step`
component. Thanks to this we can easily reuse this component in other
places and change slightly the UI- eg in the `Unmint` page we need 2
badges but the previous implementation of the `Timeline` components
renders only one badge.