Update bitcoin utils in threshold-ts lib
Add a new function that checks whether the address is `P2SH` or `P2WSH`.
For redemption(unminting) the tBTC system allows for `P2PKH`, `P2WPKH`,
`P2SH` or `P2WSH` address so we will use this function in unminting form
validation. Here we also update the `network` type in
`isValidBtcAddress`- use the type from `threshold-ts` lib instead of the
`Network` from `bitcoin-address-validation` pacakge because we operate
in the dapp on our types. In that case the external type from
`bitcoin-address-validation` package is used only in one place and there
is no need to cast type from `BitcoinNetwork` to `Network` in other
places.