Update the Redemption flow
Check if there is a wallet in network that has enough BTC to handle a
redemption request. If such a wallet exists we trigger the submit
function and pass the wallet data to this function. Otherewise the
function from the threshold-ts lib throws an error and we set an error
message for the `amount` field.