Add methods that fetch redemption events and data
To fetch the redemption details data we need query events to find the
redemption request and verify whether the redemption was handled
successfully or timed out.

Here we also noticed a bug in the `ethers.js` lib- the `ethers.js` lib
encodes the `bytesX` param in the wrong way. It uses the left-padded
rule but based on the Solidity docs it should be a sequence of bytes in
X padded with trailing zero-bytes to a length of 32 bytes(right-padded).
See
https://docs.soliditylang.org/en/v0.8.17/abi-spec.html#formal-specification-of-the-encoding
Consider this wallet public key hash
`0x03B74D6893AD46DFDD01B9E0E3B3385F4FCE2D1E`:
- `ethers.js` returns
  `0x00000000000000000000000003b74d6893ad46dfdd01b9e0e3b3385f4fce2d1e`
- should be:
  `0x03b74d6893ad46dfdd01b9e0e3b3385f4fce2d1e000000000000000000000000`

In that case, in methods that fetch the past events by indexed param
which has `bytesX` type(`RedemptionsCompleted`, `RedemptionRequested`,
`RedemptionTimedOut`) we build the filter topics manually.