Update `useFetchRedemptionDetails` hook
Compare correctly the `scriptPubKey` data from Bitcoin transacion
outputs with the redeemer output script from Ethereum event. The
redeemer otput script from the Ethereum event is prepended by the script
length encoded as a Bitcoin varint but the `scriptPubKey` is in a raw
format so we need to prefix the output script bytes buffer with 0x and
its own length.

Here we also hardcoded data in the `UnmintDetails` component for a
redemption that was already handled successfully- for testing purposes.