Temporary fix for the wallet connect v2 modal
The wallet connect v2 modal had some issues when we keep our dApp modal with the
connction status in the background. It was unable to use a search bar to search
for specific wallet and we couldn't use the scrollbar there. That means that it
was impossible to use last 8 wallets.

As a temporary fix we close our dApp modal when the walletconnect wallet opens.
Unfortunately this will result in no success/error modal when successfuly
connecting or disconnecting the walletconnect.