Fix bug when connecting on a wrong chain
When we were connecting on the wrong chain through WalletConnect there was a bug
that prevented to show the modal again after refresh. For example I've tried to
connect through MetaMask mobile using WalletConnect and Mainet chain on
localhost with goerli contracts. The page refreshed and the window did not for
WalletConnect did not appear again unless I've changed the chain in MetaMask. I
assume the user was saved and the problem was that the provider was not
diconnected.

To fix that I have to change the behaviour for chainc hanged in `App.tsx`
effect.