Refactor `buildRedemptionDetailsLink`
Use builder pattern to build the redemption details link. Also keep the
`buildRedemptionDetailsLink` function to not break the current API. We
need more generic approach to build link because in
`RedemptionRequested` evnet we do not have `btcAddress` but
`redeemerOutputScript` in expected format. Also in event we can find
`walletPublicKeyHash` not `walletPublicKey` - using builder we can build
link from `walletPublicKey` or `walletPublicKeyHash` and `btcAddress` or
`redeemerOutpuScript` and the builder class is responsible for
converting params to correct format.