Improve UX in redemption details page
Redirect to a success step when the redemption is complete. Here we
subscribe to the `RedemptionsCompleted` event and check if this event is
related to a given redemption request by checking the wallet public key
hash. If yes, we are looking for a Bitcoin transfer to a given redeemer
output script in a given bitcoin tx hash (we can get this transaction
hash from the `RedemptionsCompleted` event). If we are able to find that
transfer it means this `RedemptionsCompleted` events is related to a
given redemption request and we should redirect user to a success step.