Fix "Cannot convert undefined or null to object"
Sometime there is an error "Cannot convert undefined or null to object"
when trying to connect to walletconnect or ledger live. To reproduce it
(before the fix in this commit of course) you should:
1. Connect through Ledger Live
2. Disconnect
3. Connect through Ledger Live again (with the same account)
4. Disconnect
5. Connect through WalletConnect (it should automatically connect with
the same account that you used while connecting to Ledger Live)
6. Disconnect
7. Connect again through Ledger Live
8. The error should appear

The similar issue was descibed here: https://github.com/WalletConnect/walletconnect-monorepo/issues/3165

This commit fixes that issue by clearing out the local storage from the
walleconnect realted data.