Move `parentPathBase` to `RouteProps`
Having `parentPathBase` in `ParentComponent` type (but not in `RouteProps`)
actually complicated more thing than it resolves. I've decided to put it in
`RouteProps` as an optional parameter. Then we assing the correct
`parentPathBase` to a specific `RouteProps` object inside `renderPageComponent`
function.

This way we don't have to pass `parentPathBase` parameter to all PageComponents
manually all the way to `SubNavigaionPillsProps` because we already have it in
`links` parameter.