Add newest version of coverage pools
Up until now we only had one version of coverage pools installed. On relesae
branch we use v1.0.0, which is related to keep token. Here we add another
version of coverage pools and call it `@threshold-network/coverage-pools` to
indicate that this one is related to T token.

We use `development` tag for both of the libs as default but on production we
should use `1.0.0` for `@keep-network/coverage-pools` and `2.0.0` for
`@threshold-network/coverage-pools`.