Fix fetching `OwnerRefreshed` event
To fetch events where a given `owner` is `oldOwner` or `newOwner` we
have to do two separate calls because if we filter events by `oldOwner`
and `newOwner` at once both params must be equal to owner - it means
that the evnet should be emitted where `oldOwner = newOwner = owner`
(ethers uses `AND` condition not `OR` under the hood). Here we fetch
`OwnerRefreshed` events by `oldOwner` and `newOwner` in a two seprate
calls and then connect them to a single array.