Fix filtering out current owner stakes
We call `rolesOf` using multicall contract for all possible staking
providers (outdated and current) of a given owner address. Multicall
returns result in the same order in which was called so we can add the
staking provider address from `possibleStakingProviders` array to a
result from multicall (the `rolesOf` fn only returns owner, authorizer
and beneficiary; it does not include the staking provider address).
Given that fact we can correctly define the current staking providers.

In previous impl, first, we filtered out `rolesOf` result to have only
those whose owner is one that we passed to this function and then we map
to a staking provider address by referring to `possibleStakingProviders`
by index, which resulted in incorrect results.