Update T coverage-pools dependency
In the `package.json` there are two different versions of `coverage-pools` used.
This is done to allow using contracts from two different cov pool versions on
the mainnet (needed for computations of the TVL-related fields) - one relates to
KEEP token cov pool (<2.0.0) and one to to T token one (>=2.0.0).
On the testnet environmet we currently don't have the T token cov pool deployed.
The latest `goerli`- tagged version of the cov-pools package (`1.1.0-goerli.0`)
contains KEEP-token related contracts. We want to stick to this version (we'll
need to update it once we switch to Sepolia). For the
`@threshold-network/coverage-pools` ideally we'd want to use a T token related
pool, which if deployed for given environment, should be present in the latest
`goerli`/`sepolia` package. If the T-related pool is not yet deployed for given
environment, KEEP-related cov-pool contracts will be used in all TVL
computations (which differs from the mainnet behavior, but is acceptable on the
test env).