Create LedgerLiveAppManager class
The `LedgerLiveAppManager` will be a core for all the logic used in the
LedgerLive app. It will contain two separate managers for ethereum and bitcoin
transactions.

For ethereum manager we also have to create a separate LedgerLiveApp signer that
implements Signer class from Ethers. It is needed so that we can pass it to our
new tbtc-v2 SDK and use it to sign transactions and communicatin with the
contracts. This way we won't have to change anything in the SDK just for the
LedgerLive implementation and we keep it clean.

In this commit only ethereum manager is created. In the future we should also
create bitcoinManager and put it here, which allow the user to send bitcoins to
a specific bitcoin address.